Improve user login experience
Before the profile wasn't much and we made it on the fly as needed. Now we want every user to have a real actual profile so we can display their (non-unique) user name with monsters. To that end there's new login logic for pages that require login:

-If the user isn't logged in, log them in
-If the user is logged in but has no profile, make them one
-Otherwise, continue as planned

This is implemented with some new base classes. Most templates now expect a user object as well.

With that in mind, there really needs to be some refactoring. This works, but it's ugly.