Refactor login logic to be more maintinable.
It's now up to the template to display some kind of login prompt instead of auto-redirecting. The login url will always redirect the user back to the login url, which checks if they have a profile. If they do, it they're redirected to home. Otherwise, a profile is created for them.